refactor(crashtracker): organize signal-safe collector
